0.1.0 修复签到功能 改正道具工厂

+++ src/main/java/cn/chahuyun/entity/UserBackpack.java
@@ -23,7 +23,7 @@ public class UserBackpack implements Serializable {
     /**
      * 用户id
      */
-    private long userId;
+    private Long userId;
     /**
      * 道具编码
      */
@@ -36,7 +36,7 @@ public class UserBackpack implements Serializable {
     public UserBackpack() {
     }
 
-    public UserBackpack(int userId, String propsCode, int propId) {
+    public UserBackpack(Long userId, String propsCode, int propId) {
         this.userId = userId;
         this.propsCode = propsCode;
         this.propId = propId;
